backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1
requests==2.32.3
anthropic==0.39.0